hub.c: Fix login on FlexHub when a HN == 0 check has been enabled
This is hardly a "fix" and more of a "hack". It also only works when a
password has been configured for the hub - interactive /password-based
login will still fail.

(The "correct" approach, according to the ADC spec, is to send 0/0/0 or
simply avoid sending any hub counts in the IDENTIFY state and sending
those only after reaching the NORMAL state. I'm not sure whether other
hubs like this approach, though.)