Send SUDP key with outgoing SCH + Handle incoming encrypted UDP packets
As usual, when the ADC spec says to add "XXXX" to the SU field of INF,
you're supposed to add "XXX0" or "XXX1". Gotta love how extension
versioning is so well documented.

Jucy actually does something with the SUDP/SUD1 flags in INF, ncdc
doesn't. The presence of the KY field in SCH is assumed to indicate SUDP
support.

Ncdc also doesn't send the SUDP/SUD1 flag in HSUP. Not sure why the spec
mentions that in the first place, but I'm not aware of any hubs doing
anything with it.

Also, YAY ENCRYPTED SEARCH RESULTS!!! ZOMFG.