Use makeheaders to generate a specific header for each .c file
This is a bit different from the previous approach where a .h file was
generated for each .c file as well, except that those .h files contained
all the exported definitions. With this new approach, the .h files are
not supposed to be included in any file besides the correspondig .c
file, and they now contain the exported definitions used from other .c
files. This is all explained in the makeheaders documentation at
http://www.hwaci.com/sw/mkhdr/makeheaders.html.

This will likely need a bit more tweaking here and there.