Programming style: typedef all structs with name_t and use those
This change is only partially because my own programming style has just
changed a bit in this direction.* The other (and more practical) reason
is that makeheaders already creates the typedefs automatically, I've
just never used them. This automatic typedef creation was a bit
problematic in makeheaders' dependency resolution. E.g. 'struct hub' was
typedeffed to put the name 'hub' has a global type, but I also use 'hub'
as a variable in various places. This conflict is not really a problem
for plain C, as types and variables live in different namespaces, but
makeheaders doesn't really know that, and considered the hub structure
and its dependencies as a dependency of db.c, which uses the variable
'hub'. That's somewhat inefficient and messy...

These changes where mostly made by some runs of sed. If there were any
errors in these changes, I suspect the compiler would have found them.
(That is not to say I haven't done any basic verification myself, but,
you know, this change affects a *lot* of code >_>)

*) Not entirely, though. I was going more in the direction of using the
old struct names directly as typedefs, without the _t. This has the same
downside as mentioned above, so doesn't really work. :-(