ui_listing: Add support for hiding rows + uit_conn: Add state filters
I'm pretty sure I introduced a bug in here somewhere, as the skip
support in ui_listing wasn't very easy. It's also a bit slower due to
the requirement of using _next() and _prev() as if it were a linked list
rather than a GSequence, but the current code only walks about 3 or 4
times the height of the ncurses window, so that's not a really huge
problem.