Improved support for group chat
Things mostly work as before, but the msg tab only keeps track of where
it needs to reply to now, rather than with whom we're actually talking.
This also means that all messages with the same reply-to user will be
grouped in the same tab, rather than all messages originating from the
same user as before. This new behaviour (supposedly) matches that of
DC++ and derivatives.

Group chat on DC still sucks balls, though.