fl_save.c: Only check the buffer size once for each fl_list_t entry
An fl_list_t entry never exceeds more than a few 100 bytes at most, so
the buffer couldn't have grown significantly. This also makes the code a
bit easier to work with, as the macros won't cause a 'return' anymore
the return value of al() does not need to be checked anymore (it can't
fail). Also increased the maximum buffer size a bit, to avoid having
GString allocate power-of-two memory when we're constantly growing just
a tiny bit above power-of-two.

These changes resulted in a 10% performance increase for in-memory
serialization. When writing to a file the bzip2 compression dominates
everything, so this improvement is only useful for partial file list
generation on ADC.