fl_save.c: Allow serializing directly to a zlib-compressed buffer
Not used yet, but this will power partial file list transfers compressed
with the ZLIG extension. Still not very useful at the moment, but I'll
fix the fl_save() stuff to serialize more than a single level for
partial lists later on, in which case zlib compression gets kinda
mandatory.

Obviously, this adds zlib as a dependency. I can't think of any insane
setup where that's a problem, especially considering ncdc is released in
a .tar.gz and gnutls also requires(?) zlib.