cc.c: Support ZLIG for partial lists on ADC
Note that ncdc now indicates support for ZLIG, but doesn't actually
handle the case when a "CSND .. ZL1" is received. This is a violation of
the spec and a bug in ncdc, but I am not aware of any client that
actually sends zlib-compressed data when we haven't explicitely asked
for it.