hub.c: Lower nfo timer to 1min + don't include cur hub in HN/HR on ADC
ADC is quite annoying with regards to when certain INF flags should be
sent to the hub. Flexhub tends to check the hub counts sent in the first
NFO, and assumes that these counts do not include the current hub (which
is, according to the spec, a correct assumption).

The timer is then used to correctly update the counts after login after
some timeout. The 1 minute timeout mimics the (measured) behaviour of
EiskaltDC++.

(If active mode is enabled, this timer isn't necessary: A second NFO is
always sent directly after login to indicate that the client is active.
This, however, is also tricky: ncdc won't be able to login to a hub that
*requires* clients to be in active mode on login. All these undocumented
restrictions are a complete chaos)