fl_load.c: Don't allow '/' in file or directory name
The ADC spec allows a '/' in a file or directory name, but only when
escaped with a '\'. I've no idea how to securely deal with such
filenames, so I'll just disallow them. The '\' character is also allowed
by the spec and should be escaped similarly. Those don't pose a problem
on unix-like systems, so I suppose we can keep allowing it. No attempt
is made to unescape filenames; it wouldn't matter.