Rewrite UDP sending code + respect local_bind setting
The new UDP sending code doesn't queue messages anymore, so search
results might get dropped if the kernel send buffer is full. I'm
assuming that this doesn't happen often, but I'll have to do some
testing in order to be sure.

This commit fixes a long-standing bug where the local_bind setting was
not used for outgoing UDP messages.