dl.c: Remove dl->{active,dlthread,deleted} and handle item deletion
The dl->active field used to indicate whether a dl item was being
requested from a user or was in the process of downloading, and the
dl->dlthread field used to indicate whether an OS thread was still alive
on the dl item.

These have now been merged into a single refcount-style active_threads.
dl->active was used to avoid requesting the same file from multiple
users, but with segmented downloading that's not what we want. Instead,
dl->allbusy is used to check whether there is still some data available
to download.

The old dl_queue_checkrm() function existed primarily because it was
possible for a user to still be noted as being in the ACT state at the
time a download has finished, and immediately deleting the dl item would
cause the user to be disconnected. I've fixed the interaction with cc.c
to ensure that a user will always move out of the ACT state before the
download is considered complete (this was fairly trivial...), so the old
dl_queue_checkrm() function was obsolete.

I've also removed dl->deleted. This field is redundant because a deleted
dl item is not present in the dl_queue hash table.