dlfile: Handle incoming file loading and bitmap saving errors
An error with saving the bitmap while downloading is easily recoverable,
so the error flag is set.

An error while loading an incoming file on startup used to be handled by
deleting the incoming file and all partially downloaded data. This is
obviously not what you want to happen if the error just happened to be a
temporary issue (e.g. permission issue). Ideally, these errors would
also be handled by setting the error flag for the affected downloads,
but there is no recovery path for such errors yet, so that's not going
to be very useful. Instead, this commit will cause ncdc to refuse to
start when loading an incoming file failed, so that the user at least
has a chance to fix the issue without losing all downloaded data.