dlfile: Ensure that threads are always block-aligned
Otherwise the hash_tth state isn't correctly updated and you get hash
failures. In order to enforce this, dl->allbusy has to indicate whether
there is still an unallocated block available, rather than simply an
unallocated chunk.

This also means that the number of sources from which a file can be
downloaded simultaneously is limited by the number of unallocated blocks
in a file, which is noticable towards the end of the download when the
last parts of a file have to be received from one user. (This effect
isn't specific to this implementation, all clients suffer from this, but
other clients might have better granularity than one block. I've no idea
how other clients allocate segments, actually...)