dlfile: Use separate 'busy' flag for dl_thread_t objects
The busy status was previously inferred from the 'allocated' count, but
it is possible that allocated drops to 0 before dlfile_recv_done() is
called. In turn, it was possible that the same thread is re-used in
dlfile_getchunk(), followed by its allocated count being reset to 0 in
dlfile_recv_done(), followed by its allocated count getting wrapped to
2^32-1 in dlfile_recv(), followed by the thread never ever getting
downloaded to completion.

Separating the busy status from the allocated count should prevent this
situation. (But I haven't tested this very thoroughly yet)