dlfile: Don't consider a files.xml.bz2 file complete on error
dl->size is compared with the size of the downloaded data (dl->have) to
determine whether a file list has been downloaded completely. But when
the other user replies with an error (e.g. slots full), dl->size is not
updated with a real size and remains 0. At that point, no data has been
downloaded yet and the file list download is considered complete.

This adds a dl->hassize field as an extra check to ensure that we have
received a valid size before considering the download as complete.