search: Swap ',' and '.' keys + minor refactoring
To be consistent with ncmpc, the , key is next and the . key previous.

I've replaced the function LUTs with just a ternary operator, since the
functions were only being used once and this is much shorter.
(Performance might be slightly worse, but I highly doubt it's
significant)

I've replaced the SEARCH_NEXT/SEARCH_PREV definitions with just a
boolean. Again, shorter, but also clearer when not used as an array
index.

The definition of the REGEX_ things shouldn't be in #if INTERFACE,
they're local in ui_util.c.