Dont freopen stderr, log to separate file handle
There shouldn't be any other code that writes to stderr, so no real need
to reopen it.

This fixes the bracketed paste mode that I broke in an earlier commit by
having the codes sent to stderr. Since stderr was mapped to the log
file, the codes would never reach the terminal and the bracketed paste
mode was simply never enabled.

This fix also makes me run into the vim issue described at reply 7 of
http://dev.yorhel.nl/ncdc/bug/66 - I'll have to look into that again.

For some reason the display of non-ASCII characters in the line editing
ui is broken, too. I'll look at that in a bit.