Disable RC4 ciphers by default from tls_priority
I can't say I'm a huge fan of referencing specific ciphers and features
in the default tls_priority setting for two reasons:

1. If the particular feature isn't supported by the GnuTLS version being
   used, it will cause GnuTLS to return with an error during connect and
   all TLS activity will fail.

2. I am not an authority in figuring out which ciphers are "secure
   enough", nor do I have any clue as to which settings may cause
   issues with compatibility. GnuTLS has a "NORMAL" setting for a
   reason.

As for (1), I think it's safe to assume every GnuTLS library has support
for RC4, considering its wide-spread use. Though if RC4 is ever removed,
ncdc will break.

For (2), it's unfortunate that the "NORMAL" setting doesn't exclude RC4
already. Likely for compatibility reasons in the context of HTTPS. All
DC clients and hubs ought to support better ciphers so it shouldn't be
an issue for us to exclude it.

Fixes bug #73: http://dev.yorhel.nl/ncdc/bug/73