tls_priority: Remove ARCFOUR-40 from default exclusion list
This cypher has been removed in GnuTLS 3.4, and thus causes connection
attempts to fail if it's still in the exclusion list. I very strongly
doubt that ARCFOUR-40 has ever been in the "NORMAL" cypher selection, so
this probably doesn't change anything.

GnuTLS 3.4 also by default excludes ARCFOUR-128 from the NORMAL
selection, but it doesn't hurt to keep that around for older versions.