Mark queued files with Q and shared with S in search and file list tabs
This is an extension of the already available "S" annotation in file list tabs
that was suggested as a TODO in uit_search.c. Now, both tabs will uniformly
display "Q" next to the file if it's currently in the download queue, and "S" if
it's already in the share.