Drop support for libgeoip, add support for libmaxminddb
The GeoIP Legacy databases are not updated anymore, so libgeoip is now
useless. GeoIP support now requires libmaxminddb instead.

The new GeoIP databases are distributed with both IPv4 and IPv6 support
in a single file, so I've merged the 'geoip_cc4' and 'geoip_cc6'
settings in a single new 'geoip_cc' setting.

There is no default database path anymore, so users will have to
manually set the path to the mmdb file with the 'geoip_cc' setting.